// src/api/models/jobOfert.ts
export interface JobOfert {
  id: number;
  recruitmentEndDate: string;
  positionName: string;
  positionLevel: string;
  employmentContract: string;
  employmentType: string;
  jobType: string;
  salaryMinimum: number;
  salaryMaximum: number;
  workDays: string;
  workStartHour: string;
  workEndHour: string;
  createdAt: string;
}